Add files via upload
This Script i used on my personal PC with one idea in mind. You finished for the day and are ready for some entertainment. The script will open a game in this case StarCraft 2, discord so you can talk with your friends and a link to youtube music to give you some good jams. Further more two dialog boxes, 1 will give you encouragement as you start and the other that will display after 2.5 hours to remind you to go to bed.